start factoring out posts module
