+++ app/routes/_actions+/create-post.tsx
@@ -5,7 +5,7 @@ import { zfd } from 'zod-form-data'
 import { db } from '#app/db.ts'
 import { fallacyDetection } from '#app/modules/fallacies/fallacy-detection-client.ts'
 import { storeFallacies } from '#app/modules/fallacies/fallacy-repository.ts'
-import { createPost } from '#app/repositories/post.ts'
+import { createPost } from '#app/modules/posts/post-service.ts'
 import { requireUserId } from '#app/utils/auth.server.ts'
 
 const postDataSchema = zfd.formData({
